+++ csaps/_base.py
@@ -1,10 +1,6 @@
 # -*- coding: utf-8 -*-
 
-"""
-Cubic spline approximation (smoothing)
-
-"""
-
+import abc
 import typing as t
 
 import numpy as np
@@ -12,48 +8,34 @@ import numpy as np
 import scipy.sparse as sp
 import scipy.sparse.linalg as la
 
-
-__version__ = '0.4.0'
-
-_BreaksDataType = t.Union[
-    # Univariate data sites
-    np.ndarray,
-
-    # Grid data sites
-    t.Union[
-        t.List[np.ndarray],
-        t.Tuple[np.ndarray, ...]
-    ]
-]
-
-_UnivariateDataType = t.Union[
-    np.ndarray,
-    t.Sequence[t.Union[int, float]]
-]
-
-_UnivariateVectorizedDataType = t.Union[
-    _UnivariateDataType,
-    t.List['_UnivariateVectorizedDataType']
-]
-
-_MultivariateData = t.Union[
-    np.ndarray,
-    t.Sequence[_UnivariateDataType]
-]
-
-_GridDataType = t.Sequence[_UnivariateDataType]
+from csaps._types import (
+    BreaksDataType,
+    UnivariateDataType,
+    UnivariateVectorizedDataType,
+    MultivariateDataType,
+    NdGridDataType,
+)
 
 
 class SplinePPForm:
     """Spline representation in PP-form
+
+    Parameters
+    ----------
+    breaks : BreaksDataType
+        Breaks values vector or list of vectors (for nd-grid)
+    coeffs : np.ndarray
+        Spline coefficients
+    dim : int
+        Spline dimension (>= 1 for nd-grid/multivariate)
     """
 
-    def __init__(self, breaks: _BreaksDataType, coeffs: np.ndarray, dim: int = 1):
+    def __init__(self, breaks: BreaksDataType, coeffs: np.ndarray, dim: int = 1):
         self.gridded = isinstance(breaks, (tuple, list))
         self.breaks = breaks
         self.coeffs = coeffs
-        self.pieces = None  # type: t.Union[int, t.Tuple[int, ...]]
-        self.order = None  # type: t.Union[int, t.Tuple[int, ...]]
+        self.pieces = None  # type: t.Optional[t.Union[int, t.Tuple[int, ...]]]
+        self.order = None  # type: t.Optional[t.Union[int, t.Tuple[int, ...]]]
 
         if self.gridded:
             self.pieces = tuple(x.size - 1 for x in breaks)
@@ -77,13 +59,16 @@ class SplinePPForm:
                  self.gridded, self.breaks, self.coeffs.shape, self.coeffs,
                  self.pieces, self.order, self.dim)
 
-    def evaluate(self, xi, shape=None):
+    def evaluate(self, xi: t.Union[UnivariateDataType, NdGridDataType],
+                 shape: t.Sequence[int] = None):
         """Evaluate spline on given data sites or grid
 
         Parameters
         ----------
-        xi: X data vector or list of vectors for multivariate spline
-        shape: tuple The shape for univariate case. It determines univariate vectorized Y data shape
+        xi : UnivariateDataType, NdGridDataType
+            X data vector or list of vectors for multivariate spline
+        shape : Sequence[int, ...]
+            tuple The shape for univariate case. It determines univariate vectorized Y data shape
         """
         if self.gridded:
             return self._grid_evaluate(xi)
@@ -150,7 +135,32 @@ class SplinePPForm:
         return yi.reshape(nsize, order='F')
 
 
-class UnivariateCubicSmoothingSpline:
+class ISmoothingSpline(abc.ABC):
+    """The interface class for smooting splines
+    """
+
+    @property
+    @abc.abstractmethod
+    def smooth(self) -> t.Union[float, t.Tuple[float, ...]]:
+        """Returns smoothing parameter
+        """
+        pass
+
+    @property
+    @abc.abstractmethod
+    def spline(self) -> SplinePPForm:
+        """Returns spline representation in PP-form
+        """
+        pass
+
+    @abc.abstractmethod
+    def __call__(self, xi: t.Union[UnivariateDataType, NdGridDataType]) -> np.ndarray:
+        """Evaluates spline on the data sites
+        """
+        pass
+
+
+class UnivariateCubicSmoothingSpline(ISmoothingSpline):
     """Univariate cubic smoothing spline
 
     Parameters
@@ -168,12 +178,12 @@ class UnivariateCubicSmoothingSpline:
     """
 
     def __init__(self,
-                 xdata: _UnivariateDataType,
-                 ydata: _UnivariateVectorizedDataType,
-                 weights: t.Optional[_UnivariateDataType] = None,
+                 xdata: UnivariateDataType,
+                 ydata: UnivariateVectorizedDataType,
+                 weights: t.Optional[UnivariateDataType] = None,
                  smooth: t.Optional[float] = None):
 
-        self._spline = None  # type: SplinePPForm
+        self._spline = None  # type: t.Optional[SplinePPForm]
         self._smooth = smooth
 
         (self._xdata,
@@ -186,7 +196,7 @@ class UnivariateCubicSmoothingSpline:
 
         self._make_spline()
 
-    def __call__(self, xi: _UnivariateDataType) -> np.ndarray:
+    def __call__(self, xi: UnivariateDataType) -> np.ndarray:
         """Evaluate the spline's approximation for given data
         """
         xi = np.asarray(xi, dtype=np.float64)
@@ -322,7 +332,7 @@ class UnivariateCubicSmoothingSpline:
         self._spline = SplinePPForm(self._xdata, coeffs, self._ydim)
 
 
-class MultivariateCubicSmoothingSpline:
+class MultivariateCubicSmoothingSpline(ISmoothingSpline):
     """Multivariate parametrized cubic smoothing spline
 
     Class implments multivariate data approximation via cubic smoothing spline with
@@ -372,9 +382,9 @@ class MultivariateCubicSmoothingSpline:
     """
 
     def __init__(self,
-                 data: _MultivariateData,
-                 tdata: t.Optional[_UnivariateDataType] = None,
-                 weights: t.Optional[_UnivariateDataType] = None,
+                 data: MultivariateDataType,
+                 tdata: t.Optional[UnivariateDataType] = None,
+                 weights: t.Optional[UnivariateDataType] = None,
                  smooth: t.Optional[float] = None):
         self._data, self._tdata = self._prepare_data(data, tdata)
 
@@ -386,20 +396,22 @@ class MultivariateCubicSmoothingSpline:
             smooth=smooth
         )
 
-    def __call__(self, ti: _UnivariateDataType):
+    def __call__(self, ti: UnivariateDataType):
         return self._univariate_spline(ti)
 
     @property
-    def t(self) -> np.ndarray:
-        return self._tdata
+    def smooth(self) -> float:
+        return self._univariate_spline.smooth
 
     @property
     def spline(self) -> SplinePPForm:
         return self._univariate_spline.spline
 
     @property
-    def smooth(self) -> float:
-        return self._univariate_spline.smooth
+    def t(self) -> np.ndarray:
+        """Returns parametric data vector
+        """
+        return self._tdata
 
     @staticmethod
     def _compute_tdata(data):
@@ -430,7 +442,7 @@ class MultivariateCubicSmoothingSpline:
         return data, tdata
 
 
-class NdGridCubicSmoothingSpline:
+class NdGridCubicSmoothingSpline(ISmoothingSpline):
     """ND-Gridded cubic smoothing spline
 
     Class implments ND-gridded data approximation via cubic smoothing spline
@@ -456,9 +468,9 @@ class NdGridCubicSmoothingSpline:
     """
 
     def __init__(self,
-                 xdata: _GridDataType,
+                 xdata: NdGridDataType,
                  ydata: np.ndarray,
-                 weights: t.Optional[t.Union[_UnivariateDataType, _GridDataType]] = None,
+                 weights: t.Optional[t.Union[UnivariateDataType, NdGridDataType]] = None,
                  smooth: t.Optional[t.Union[float, t.Sequence[float]]] = None):
         (self._xdata,
          self._ydata,
@@ -466,7 +478,7 @@ class NdGridCubicSmoothingSpline:
          self._smooth) = self._prepare_data(xdata, ydata, weights, smooth)
 
         self._ndim = len(self._xdata)
-        self._spline = None  # type: SplinePPForm
+        self._spline = None  # type: t.Optional[SplinePPForm]
 
         self._make_spline()
 
@@ -539,7 +551,7 @@ class NdGridCubicSmoothingSpline:
 
         return xdata, ydata, weights, smooth
 
-    def __call__(self, xi: _GridDataType) -> np.ndarray:
+    def __call__(self, xi: NdGridDataType) -> np.ndarray:
         xi = self._prepare_grid_vectors(xi, 'xi')
 
         if len(xi) != self._ndim: