+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.5.0
+
+* Reorganize the project to package-based structure
+* Add the interface class for all smoothing spline classes
+
 # v0.4.2
 
 * FIX: "smooth" value is 0.0 was not used